Template improvements
- Flags: templatise enum type, remove s32 casts.
- resources: Use templated constexpr bool to
static_assert only after instantiation.
- Replace `getString` etc with templated types.
- Rename `toStr` to `toString`.
- Make Colour::lerp more safe.
- Rename `name` to `id` for all resources.
- Rename `Typename()` to `typeName()`.
- Extract `demangle()` from `typeName()`.
- Use generic `ObjMap<Key, Obj, MapType>` with
`if constexpr` to support both `unordered_map`
and `map`.
- Have `resources::get<T>()` return `const&`.
- Remove `m_tint` from `Model`.
- Store `Model const*` in `Prop`.

+++ le3d/src/engine/contextImpl.cpp
@@ -246,16 +246,16 @@ void contextImpl::clearFlags(context::ClearFlags flags, Colour colour)
 	{
 		cxChk();
 		GLbitfield glFlags = 0;
-		if (flags.isSet(s32(context::ClearFlag::ColorBuffer)))
+		if (flags.isSet(context::ClearFlag::ColorBuffer))
 		{
 			glChk(glClearColor(colour.r.toF32(), colour.g.toF32(), colour.b.toF32(), colour.a.toF32()));
 			glFlags |= GL_COLOR_BUFFER_BIT;
 		}
-		if (flags.isSet(s32(context::ClearFlag::DepthBuffer)))
+		if (flags.isSet(context::ClearFlag::DepthBuffer))
 		{
 			glFlags |= GL_DEPTH_BUFFER_BIT;
 		}
-		if (flags.isSet(s32(context::ClearFlag::StencilBuffer)))
+		if (flags.isSet(context::ClearFlag::StencilBuffer))
 		{
 			glFlags |= GL_STENCIL_BUFFER_BIT;
 		}
